Pass SerializationConfiguration as param to UseSerialization function
