+++ commercetools.Sdk/Tests/commercetools.Sdk.Serialization.Tests/SerializationFixture.cs
@@ -19,5 +19,16 @@ namespace commercetools.Sdk.Serialization.Tests
         }
 
         public ISerializerService SerializerService { get; private set; }
+
+        public ISerializerService BuildSerializerServiceWithConfig(
+            SerializationConfiguration config = null)
+        {
+            var services = new ServiceCollection();
+            services.UseRegistration();
+            services.UseDomain();
+            services.UseSerialization(config ?? SerializationConfiguration.DefaultConfig);
+            var serviceProvider = services.BuildServiceProvider();
+            return serviceProvider.GetService<ISerializerService>();
+        }
     }
 }
