+++ commercetools.Sdk/commercetools.Sdk.Serialization/SerializationConfiguration.cs
@@ -2,6 +2,7 @@ namespace commercetools.Sdk.Serialization
 {
     public class SerializationConfiguration : ISerializationConfiguration
     {
+        public static readonly SerializationConfiguration DefaultConfig = new SerializationConfiguration();
         public bool DeserializeDateAttributesAsString { get; set; }
         
         public bool DeserializeDateTimeAttributesAsString { get; set; }
